feat(store): implement devtools integration and refactor store structure
- Add devtools integration for store management with functions to register, unregister, and notify state changes.
- Refactor store module to separate concerns into distinct files: create-store, mapping, persisted, plugins, registry, types, utils, and watch.
- Introduce mapping helpers for state, getters, and actions to simplify component integration.
- Implement persistence functionality for stores using localStorage.
- Add watch functionality to observe state changes with customizable options.
- Enhance tests to cover new features including mapGetters and watchStore.
